+++ src/ZoneTree/Core/IncrementalIdProvider.cs
@@ -2,13 +2,13 @@
 
 public class IncrementalIdProvider : IIncrementalIdProvider
 {
-    int lastId = 0;
-    public int NextId()
+    long lastId = 0;
+    public long NextId()
     {
         return Interlocked.Increment(ref lastId);
     }
 
-    public void SetNextId(int id)
+    public void SetNextId(long id)
     {
         Interlocked.Exchange(ref lastId, id - 1);
     }
